refact: separate s3 registry file from `local.go`
